backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
billiard==4.2.4
cachetools==6.2.6
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==47.0.0
deprecation==2.1.0
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
django-ratelimit==4.1.0
djangorestframework==3.17.1
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
iniconfig==2.3.0
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
packaging==26.2
pluggy==1.6.0
postgrest==2.29.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.1.0
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
realtime==2.29.0
redis==7.4.0
requests==2.33.1
rich==14.3.4
six==1.17.0
sqlparse==0.5.5
storage3==2.29.0
StrEnum==0.4.15
strictyaml==1.7.3
stripe==15.1.0
supabase==2.29.0
supabase-auth==2.29.0
supabase-functions==2.29.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
vine==5.1.0
wcwidth==0.6.0
websockets==15.0.1
whitenoise==6.12.0
yarl==1.23.0
zstandard==0.25.0
